Update training demo structure
